+++ credentials.sh
@@ -20,7 +20,7 @@
 
 set -o pipefail
 
-VERSION="1.9.11"
+VERSION="1.9.12"
 TITLE="CloudVPS Boss Credentials Config ${VERSION}"
 
 if [[ ${DEBUG} == "1" ]]; then